backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
google-cloud-bigquery==3.14.1
google-auth==2.25.2
azure-purview-catalog==1.0.0b4
azure-identity==1.15.0
azure-core==1.29.5
schedule==1.2.0
apscheduler==3.10.4


